connect db

+++ src/chat/chat_http/chat_dto.go
@@ -6,7 +6,7 @@ type ChatDto struct {
 	ID          uint64    `json:"id"`
 	Name        string    `json:"name"`
 	LastMessage uint64    `json:"lastMessage"`
-	Subscribers []uint64  `json:"subscribers"`
+	Subscribers []uint8   `json:"subscribers"`
 	CreatedAt   time.Time `json:"createdAt"`
 	UpdatedAt   time.Time `json:"updatedAt"`
 }
